GUI for undertaking overtake
*Notes*
Tkinter comes with most python versions, but in some it is called "Tkinter" and in others "tkinter". Try changing the capitalization if the module can't be found. 

Due to some complexities of Tkinter, we need threads to allow the UI to be responsive and the serial communication to run simultaneously.